wp-now: split create-server run-cli logic
